UltimosCambios 07/04

+++ macowins/src/main/java/Grupo4/Negocio.java
@@ -12,7 +12,7 @@ public class Negocio
 	List <Venta> ventas=new ArrayList<>();
 	
 	
-    public float precioFinal(Prenda unaPrenda){
+    public double precioFinal(Prenda unaPrenda){
     		return unaPrenda.getPrecioFinalPrenda();
     }
 	public void realizarVenta(Prenda unaPrenda,int cantidad){
@@ -23,11 +23,11 @@ public class Negocio
     private void registrarVenta(Venta unaVenta){
     	ventas.add(unaVenta);
     }
-    public float gananciasDelDia(String fecha){ //Formato Fecha dd/m/aaaa
+    public double gananciasDelDia(String fecha){ //Formato Fecha dd/m/aaaa
     	List <Venta> ventasDelDia=new ArrayList<>();
-    	float ganancia;
+    	double ganancia;
     	ventasDelDia=ventas.stream().filter(unaVenta->unaVenta.getFecha().equals(fecha)).collect(Collectors.toList());
-    	ganancia=(float) ventasDelDia.stream().mapToDouble(unaVenta->(unaVenta.getPrecioFinal())).sum();
+    	ganancia= ventasDelDia.stream().mapToDouble(unaVenta->(unaVenta.getPrecioFinal())).sum();
     	return ganancia;
     }
     private String calcularFecha(){
